+++ pkg/deploy/devfileregistry/init_test.go
@@ -9,7 +9,7 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-package registry
+package devfileregistry
 
 import "github.com/eclipse-che/che-operator/pkg/deploy"
