+++ pkg/deploy/pluginregistry/pluginregistry.go
@@ -9,7 +9,7 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-package registry
+package pluginregistry
 
 import (
 	"strings"
