+++ pkg/deploy/pluginregistry/pluginregistry_configmap.go
@@ -9,7 +9,7 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-package registry
+package pluginregistry
 
 import (
 	"encoding/json"
