+++ pkg/deploy/pluginregistry/pluginregistry_deployment_test.go
@@ -9,7 +9,7 @@
 // Contributors:
 //   Red Hat, Inc. - initial API and implementation
 //
-package registry
+package pluginregistry
 
 import (
 	"os"
