+++ pkg/deploy/registry/registry_deployment.go
